// src/features/projects/hooks/use-files.ts
import { useMutation, useQuery } from "convex/react";
import { Id } from "../../../../convex/_generated/dataModel";
import { api } from "../../../../convex/_generated/api";

export const useFile = (fileId: Id<"files"> | null) => {
  return useQuery(api.files.getFile, fileId ? { id: fileId } : "skip");
};

export const useFilePath = (fileId: Id<"files"> | null) => {
  return useQuery(api.files.getFilePath, fileId ? { id: fileId } : "skip");
};

export const useUpdateFile = () => {
  return useMutation(api.files.updateFile);
};

export const useCreateFile = () => {
  return useMutation(api.files.createFile);

  //TODO: ADD OPTIMISTIC UPDATES
};

export const useCreateFolder = () => {
  return useMutation(api.files.createFolder);

  //TODO: ADD OPTIMISTIC UPDATES
};

export const useRenameFile = () => {
  return useMutation(api.files.renameFile);

  //TODO: ADD OPTIMISTIC UPDATES
};

export const useDeleteFile = () => {
  return useMutation(api.files.deleteFile);

  //TODO: ADD OPTIMISTIC UPDATES
};

export const useFolderContents = ({
  projectId,
  parentId,
  enabled = true,
}: {
  projectId: Id<"projects">;
  parentId?: Id<"files">;
  enabled?: boolean;
}) => {
  return useQuery(
    api.files.getFolderContents,
    enabled ? { projectId, parentId } : "skip",
  );
};
